🔧 Functions not expected for use marked as private (#38)
Wimsey has a lot of functions under the hood for its own use, like "collect_all_tests", or "parse_config" (eg). These aren't designed for actual use, so I'm marking them as private here. The goal is, that this will expose a consistent API with a lot of guarantees.

Part of Wimsey's whole deal is that it's aggresively simple to use, so I think guaranteeing a really stable API here, that people don't have to think about it changing under them, is a big benefit.

Reviewed-on: https://codeberg.org/benrutter/wimsey/pulls/38
Co-authored-by: ben <benrutter@posteo.net>
Co-committed-by: ben <benrutter@posteo.net>